GKE: Replace deprecated region with new location parameter
Location accepts both a region or zone and then respectively causes
a regional or zonal cluster to be started.

We're aiming for regional clusters only.

This commit is both to remove the deprecation warnings but also
in preparation of removing the default node pool and replacing it
with a node pool with autoscaling enabled.