GKE: Replace default with separately managed node pool
This commit disables the default node pool created with each cluster
and replaces it with a separetely managed node pool.

This enables autoscaling and switches to min, max input variables
to fix #6.

This also enables creating additional node pools, although it
requires using `_modules/gke/node_pools` which should be considered
private. So this relates to #8 but does not fix it.